+++ projectZoo/src/pr1/Event.java
@@ -10,7 +10,7 @@ public class Event extends Main {
     static String name;
     static String location;
     static String type;
-    Human h = new Human();
+    Person h = new Person();
 
     Event(String name, String location, String type) {
         this.name = name;
@@ -18,14 +18,9 @@ public class Event extends Main {
         this.type = type;
     }
 
-    void theFirstStage(){
+    void theFirstStage() {
         System.out.println("AAAAANDDD NOOOOOOWWWW we're nearly going to start our program! Please, come and pass" +
                 "the registration and who knows, maybe you we'll even be able to see the show!");
-        startRegistration();
-    }
-
-    void startRegistration(){
-        h.setHumans();
     }
 
 }
