+++ projectZoo/src/pr1/Person.java
@@ -1,0 +1,18 @@
+package pr1;
+
+import java.util.*;
+
+public class Human extends Main {
+    String name;
+    Date dateOfBirth;
+    String cityOfBirth;
+    String countryOfBirth;
+    int age;
+
+    Date date = new Date();
+    RandomQuantity random = new RandomQuantity();
+    Human(){}
+    Human(String name, int age){}
+
+    
+}
